fix: ojコマンドですでにファイルが存在する場合はテンプレートのコピーをしない

+++ sh/oj.sh
@@ -9,4 +9,10 @@ oj d $1
 # ファイルを作成する
 array=( `echo $1 | tr -s '/' ' '`)
 last_index=`expr ${#array[@]} - 1`
-cp $PWD/sh/template.php "${array[${last_index}]}.php"
+
+# テンプレートのコピー
+if [ -e "${array[${last_index}]}.php" ]; then
+  echo "ファイルが存在するためテンプレートのコピーをスキップしました"
+else
+  cp $PWD/sh/template.php "${array[${last_index}]}.php"
+fi
